+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/domain/calculated_measurements.py
