+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/domain/calculated_measurements_factory.py
@@ -1,10 +1,10 @@
 from uuid import UUID
 
-from geh_common.domain.types import MeteringPointType
+from geh_common.domain.types import MeteringPointType, OrchestrationType
 from pyspark.sql import Column, DataFrame
 from pyspark.sql import functions as F
 
-from geh_calculated_measurements.electrical_heating.domain.calculated_measurements import (
+from geh_calculated_measurements.common.domain.calculated_measurements import (
     CalculatedMeasurements,
 )
 from geh_calculated_measurements.electrical_heating.domain.column_names import ColumnNames
@@ -13,13 +13,13 @@ from geh_calculated_measurements.electrical_heating.domain.column_names import C
 def create(
     measurements: DataFrame,
     orchestration_instance_id: UUID,
-    orchestration_type: str,
+    orchestration_type: OrchestrationType,
     metering_point_type: MeteringPointType,
 ) -> CalculatedMeasurements:
     df = measurements.withColumns(
         {
             ColumnNames.orchestration_instance_id: F.lit(str(orchestration_instance_id)),
-            ColumnNames.orchestration_type: F.lit(orchestration_type),
+            ColumnNames.orchestration_type: F.lit(orchestration_type.value),
             ColumnNames.metering_point_type: F.lit(metering_point_type.value),
             ColumnNames.transaction_creation_datetime: F.current_timestamp(),
             ColumnNames.transaction_id: _add_transaction_id(),
