+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/infrastructure/calculated_measurements/database_definitions.py
@@ -1,0 +1,3 @@
+class CalculatedMeasurementsInternalDatabaseDefinition:
+    DATABASE_NAME = "measurements_calculated_internal"
+    MEASUREMENTS_NAME = "measurements"
