--- source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/infrastructure/measurements/calculated_measurements/database_definitions.py
@@ -1,3 +1,0 @@
-class CalculatedMeasurementsDatabaseDefinition:
-    DATABASE_NAME = "measurements_calculated"
-    MEASUREMENTS_NAME = "measurements"
