+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/infrastructure/measurements/repository.py
@@ -1,11 +1,5 @@
 from pyspark.sql import DataFrame, SparkSession
 
-from geh_calculated_measurements.electrical_heating.domain.calculated_measurements import (
-    CalculatedMeasurements,
-)
-from geh_calculated_measurements.electrical_heating.infrastructure.measurements.calculated_measurements.database_definitions import (
-    CalculatedMeasurementsDatabaseDefinition,
-)
 from geh_calculated_measurements.electrical_heating.infrastructure.measurements.measurements_gold.database_definitions import (
     MeasurementsGoldDatabaseDefinition,
 )
@@ -22,25 +16,7 @@ class Repository:
         catalog_name: str | None = None,
     ) -> None:
         self._spark = spark
-        self._calculated_measurements_database_name = CalculatedMeasurementsDatabaseDefinition.DATABASE_NAME
-        self._calculated_measurements_table_name = CalculatedMeasurementsDatabaseDefinition.MEASUREMENTS_NAME
         self._catalog_name = catalog_name
-        if self._catalog_name:
-            self._calculated_measurements_full_table_path = f"{self._catalog_name}.{self._calculated_measurements_database_name}.{self._calculated_measurements_table_name}"
-        else:
-            self._calculated_measurements_full_table_path = (
-                f"{self._calculated_measurements_database_name}.{self._calculated_measurements_table_name}"
-            )
-
-    def write_calculated_measurements(
-        self, calculated_measurements: CalculatedMeasurements, write_mode: str = "append"
-    ) -> None:
-        calculated_measurements.df.write.format("delta").mode(write_mode).saveAsTable(
-            self._calculated_measurements_full_table_path
-        )
-
-    def read_calculated_measurements(self) -> CalculatedMeasurements:
-        return CalculatedMeasurements(self._spark.read.table(self._calculated_measurements_full_table_path))
 
     def read_time_series_points(self) -> TimeSeriesPoints:
         # TODO: the table does not yet exist in the database
